Move Slots implementation to new file, don't reexport everything
